Add a report of failed tests to the end of run-tests.py using --jerry-test-suite
Previously the user had to scroll back to each block, to check the exact test failed, which is a bit tedious.
This PR adds a final report of them, indicating the failed block or a text, or a `'No failing tests! Good job!'` if none returned with error.

JerryScript-DCO-1.0-Signed-off-by: Bela Toth tbela@inf.u-szeged.hu